Break out windows/unix specific pipe creation.

+++ sdk/nodejs/cmd/pulumi-language-nodejs/proxy_unix.go
@@ -29,35 +29,15 @@ import (
 
 func createPipes() (pipes, error) {
 	dir, err := ioutil.TempDir("", "pulumi-node-pipes")
+	if err != nil {
+		return nil, err
+	}
+
 	invokeReqPath, invokeResPath := path.Join(dir, "invoke_req"), path.Join(dir, "invoke_res")
-
-	if err != nil {
-		return nil, err
-	}
-	if err := syscall.Mkfifo(path.Join(dir, "invoke_req"), 0600); err != nil {
-		logging.V(10).Infof("createPipes: Received error opening request pipe: %s\n", err)
-		return nil, err
-	}
-	if err := syscall.Mkfifo(path.Join(dir, "invoke_res"), 0600); err != nil {
-		logging.V(10).Infof("createPipes: Received error opening result pipe: %s\n", err)
-		return nil, err
-	}
-
-	invokeReqPipe, err := os.OpenFile(invokeReqPath, os.O_RDONLY, 0)
-	if err != nil {
-		return nil, err
-	}
-	invokeResPipe, err := os.OpenFile(invokeResPath, os.O_WRONLY, 0)
-	if err != nil {
-		return nil, err
-	}
-
 	return &unixPipes{
 		dir:     dir,
 		reqPath: invokeReqPath,
-		reqPipe: invokeReqPipe,
 		resPath: invokeResPath,
-		resPipe: invokeResPipe,
 	}, nil
 }
 
@@ -79,12 +59,42 @@ func (p *unixPipes) writer() io.Writer {
 	return p.resPipe
 }
 
-func (p *unixPipes) shutdown() {
-	contract.IgnoreClose(p.reqPipe)
-	contract.IgnoreClose(p.resPipe)
+func (p *unixPipes) connect() error {
+	if err := syscall.Mkfifo(path.Join(p.dir, "invoke_req"), 0600); err != nil {
+		logging.V(10).Infof("createPipes: Received error opening request pipe: %s\n", err)
+		return err
+	}
 
-	contract.IgnoreError(os.Remove(p.reqPath))
-	contract.IgnoreError(os.Remove(p.resPath))
+	if err := syscall.Mkfifo(path.Join(p.dir, "invoke_res"), 0600); err != nil {
+		logging.V(10).Infof("createPipes: Received error opening result pipe: %s\n", err)
+		return err
+	}
+
+	invokeReqPipe, err := os.OpenFile(p.reqPath, os.O_RDONLY, 0)
+	if err != nil {
+		return err
+	}
+	p.reqPipe = invokeReqPipe
+
+	invokeResPipe, err := os.OpenFile(p.resPath, os.O_WRONLY, 0)
+	if err != nil {
+		return err
+	}
+	p.resPipe = invokeResPipe
+
+	return nil
+}
+
+func (p *unixPipes) shutdown() {
+	if p.reqPipe != nil {
+		contract.IgnoreClose(p.reqPipe)
+		contract.IgnoreError(os.Remove(p.reqPath))
+	}
+
+	if p.resPipe != nil {
+		contract.IgnoreClose(p.resPipe)
+		contract.IgnoreError(os.Remove(p.resPath))
+	}
 
 	contract.IgnoreError(os.Remove(p.dir))
 }
